Altered rotation force calculation, minor performance improvements, lua error fix.
Changed the multiplier for HP/T from being linear to being (x^1.5)/5. I may alter this more in the future as there are a lot of other factors affecting rotation speed, which might be creating undesirable effects. In any case with this change 30 HP/T vehicles I tested can rotate roughly twice as fast as 10HP/T vehicles now, whereas before the difference was very minor.

Fixed a lua error caused by a nil value in the hud code. This error would occur when looking at a daktank vehicle that hadn't fully initialized yet. Which meant it was incredibly frequent.

I previously had noticed the checkspherical function was taking up a pretty substantial amount of execution time. I couldn't think of any way to effectively optimize the function itself, but I realized it was being called most frequently in gearbox (every tick). There's no reason it needs to be called that frequently, so I moved it into the slowthink if statement that was already there.