Selftbl on (almost) all components, began cleaning up and optimizing tankcore
Applied the selftbl optimization to all components that still had double digit execution times (ms) over 20 seconds in my testing. Individually these don't contribute much but it was easy to make this change to all of them, and all-together it should be another decent dent in optimizing this mess.

Started making a dent into making tankcore more readable, and better optimized. A lot of what I've changed so far is only run once, so it won't have much of an impact yet.