Minor optimization of ammo crates
Moved checkspherical into the slowthink if statement in Ammo crates' think function. Changed the stat-setting thing to split the dakname string once, and store it, instead of doing it repeatedly for each elseif.

This roughly halved the execution time, but it wasn't amazingly high in the first place (at least not compared to gearboxes), so this won't be a big difference. It might have a slightly bigger impact on things like artillery emplacements, where you're likely to have a lot of ammo crates.