Minor balance tweaks and cleanup
Towed guns are now 0.1 flanking mult rather than 0.13, and vehicles now have a minimum of 0.15 instead of 0.1. So towed guns actually have an advantage over powered vehicles instead of just being a direct downgrade.

Removed the invehiclebounds check, as it didn't work properly, and only served as a means for people in-the-know to circumvent other more reasonable checks. The vast majority of designs should be unaffected by this.

Did a little bit of clean-up while I was looking around. Removed some commented out code, variables/code that were un-used, and cut down a little bit of redundancy.