Tank Core Optimizations
Difficult to summarize. Mostly just the same thing I've been doing everywhere else (selfTbl, fixing obviously bad code, reducing redundant work).

The one thing of note is that I did comment out a piece of code meant to detect crew clipping. It was fairly obvious it didn't work, so it's not worth running. But it might be worth fixing and re-implementing in the future.

Most of the file still sucks, I just realized dakota conveniently marked the most intensive/frequently running part with comments, and optimized that.