TankCore initial scan optimization
Optimizing tankcore again, very shortly after I said I wouldn't do that, because someone on Valen complained about the lag-spike when tanks spawn.

Most of the lag comes from the functions it calls (Particularly the armour/penetration traces in the average armour calculation) rather than the tankcore code itself. So I've done a fair bit of work on improving those. I've made a dent in it but not too substantial of one (0.75s execution time to 0.65s). The bright side of this is that it should improve general performance as well, because these functions are used for ballistics too.

It's going to be difficult to improve it too much as a lot of the lag is probably coming from the traces themselves (Which I can't really do much to help with).