Fixed Ballistics functions relying on Hook that didn't exist.
Removed (almost) every call to the "DakTankDamageCheck" hook, which does not exist.

In most cases this results in no change (As dakota is weird and checks if the result is not equal to false for most of these). However the functions ContHEAT, ContSpall, and ContExp previously did not work properly, as they exited early if they didn't receive true from that (non-existent) hook.

The biggest impact of this change will be that HEAT and HEAT-Derived munitions will now be viable again. Internal damage should also be somewhat increased overall as well, and HE/APHE may be slightly better against vehicles.