Misc Optimizations / Print cleanup
Removed a few prints I forgot to get rid of.

Marginally improved the optimization of a lot of systems used for tank-core armour calculations.

There variance in execution times for the tank-core's scan is far higher than the amount I save with each change, so it's hard to tell how much progress I've made. I'd estimate I've reduced it by about 0.05s, which isn't very much. The bright side is quite a few of the optimizations I made should improve the performance of other ballistics-related functions as well, so hopefully that'll justify the time I spent on this.