Further Autotread Improvements & Optimization
Further reduced the amount of redundant work being done in Autotread gearboxes, particularly in the wheel loops. Also improved the efficiency of a lot of the vector calculations by reducing the number of un-necessary vector objects being created. There's still quite a bit more I can do to improve that so I've left TODO notes for a future session.

Made AbsorbForceFinal discard changes in wheel height if the wheel is in the air, or was in the air during the last tick. This should eliminate extreme change values caused by traces starting on the underside of a displacement, or moving on/off of a ledge.
Also made AbsorbForceFinal use the average of its current calculated value, and its previous value, in order to smooth out changes and reduce the impact of any remaining extreme spikes/instability.

These 2 changes should hopefully eliminate most if not all of the remaining issues with steep slopes, and other terrain that autotread doesn't like. You shouldn't be able to rocket up a slope at all now, even if hitting it at high-speed/intentionally trying to launch yourself up it. You should no longer be able to start violently spinning or flying for the same reasons.