Partially standardize weapon definitions
The different weapon types defined in dak_tegun now fully use this new system, and most of the code relating to weapon types in the Spawner tool now uses this as well. dak_teautogun and dak_temachinegun still need to be updated to use this, as well as the remaining redundant code in the Spawner tool and anything else that might hardcode weapon stats.